requirements.txt
boto3
opencv-python==4.11.0.86
numpy==1.26.1
pillow==10.4.0 
torch

